refactor/EV-778/Refactoring details files and folders
