+++ lib/details/widgets/line_chart_list_view_buttons.dart
@@ -1,0 +1,39 @@
+import 'package:crypto/details/widgets/line_chart_button_widget.dart';
+import 'package:flutter/widgets.dart';
+
+class LineChartListViewButtons extends StatelessWidget {
+  const LineChartListViewButtons({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      margin: const EdgeInsets.only(bottom: 15),
+      height: 35,
+      child: ListView(
+        scrollDirection: Axis.horizontal,
+        children: [
+          LineChartButtonWidget(
+            dayTitle: '5D',
+            titleIndex: 0,
+          ),
+          LineChartButtonWidget(
+            dayTitle: '15D',
+            titleIndex: 1,
+          ),
+          LineChartButtonWidget(
+            dayTitle: '30D',
+            titleIndex: 2,
+          ),
+          LineChartButtonWidget(
+            dayTitle: '45D',
+            titleIndex: 3,
+          ),
+          LineChartButtonWidget(
+            dayTitle: '90D',
+            titleIndex: 4,
+          ),
+        ],
+      ),
+    );
+  }
+}
