+++ lib/shared/repositories/crypto_list_repository.dart
@@ -33,7 +33,7 @@ class CryptoListRepository {
         "percent_change_last_30_days": -0.8908177605434218,
         "percent_change_last_50_days": 0.8932175668944231,
       },
-      marketPriceUpnDown: MarketPriceUpandDown,
+      marketPriceUpnDown: marketPriceUpandDown,
       userBalance: 7996.5,
       exchange: Decimal.parse('0.0000095'),
     ),
@@ -58,7 +58,7 @@ class CryptoListRepository {
         "history_price_last_30_days": 8063.22,
         "history_price_last_50_days": 7931.12,
       },
-      marketPriceUpnDown: MarketPriceUpandDown,
+      marketPriceUpnDown: marketPriceUpandDown,
     ),
     CryptoListModel(
       shortName: 'LTC',
@@ -81,7 +81,7 @@ class CryptoListRepository {
         "history_price_last_30_days": 517.42,
         "history_price_last_50_days": 207.22,
       },
-      marketPriceUpnDown: MarketPriceUpandDown,
+      marketPriceUpnDown: marketPriceUpandDown,
     ),
     CryptoListModel(
       shortName: 'DOGE',
@@ -104,7 +104,7 @@ class CryptoListRepository {
         "history_price_last_30_days": 0.20,
         "history_price_last_50_days": 0.15,
       },
-      marketPriceUpnDown: MarketPriceUpandDown,
+      marketPriceUpnDown: marketPriceUpandDown,
     ),
     CryptoListModel(
       shortName: 'ADA',
@@ -127,85 +127,7 @@ class CryptoListRepository {
         "history_price_last_30_days": 3.2,
         "history_price_last_50_days": 4,
       },
-      marketPriceUpnDown: {
-        "price_change_last_5_days": [
-          [0, 6],
-          [1.4, 2],
-          [2.4, 3.9],
-          [3, 2.5],
-          [4, 5],
-          [4.3, 5.1],
-          [5, 3.5],
-          [5.5, 5.5],
-          [6, 7.5],
-          [7.5, 6],
-          [8, 9.5],
-          [9, 10.2],
-          [9.3, 10.5],
-          [10, 12.4],
-        ],
-        "price_change_last_10_days": [
-          [0, 9],
-          [2.4, 3.9],
-          [4, 5],
-          [4.3, 5.1],
-          [5, 3.5],
-          [5.5, 5.5],
-          [6, 7.5],
-          [7.5, 6],
-          [8.5, 9.5],
-          [9, 10.2],
-          [9.3, 8.5],
-          [10, 9.4],
-        ],
-        "price_change_last_15_days": [
-          [0, 5],
-          [2.4, 3.9],
-          [4, 5],
-          [4.3, 5.1],
-          [5, 3.5],
-          [5.5, 5.5],
-          [6, 7.5],
-          [7.5, 6],
-          [8.5, 9.5],
-          [9, 10.2],
-          [9.3, 10.5],
-          [10, 9.4],
-        ],
-        "price_change_last_30_days": [
-          [0, 3.4],
-          [1, 4.9],
-          [2, 5.8],
-          [3, 7],
-          [3.5, 8],
-          [4, 9.5],
-          [5, 5],
-          [5.3, 7.5],
-          [5.7, 9.5],
-          [6.6, 7.5],
-          [7, 11],
-          [8, 8],
-          [8.5, 10],
-          [9, 12.2],
-          [10, 5],
-        ],
-        "price_change_last_50_days": [
-          [0, 5],
-          [2.4, 5.9],
-          [3, 4],
-          [4, 5],
-          [4.3, 3.1],
-          [5, 2],
-          [5.5, 5.5],
-          [6, 1.5],
-          [7, 6],
-          [8, 7],
-          [8.5, 10],
-          [8.8, 11],
-          [9, 10.5],
-          [10, 8.4],
-        ],
-      },
+      marketPriceUpnDown: marketPriceUpandDown,
     ),
     CryptoListModel(
       shortName: 'SOL',
@@ -228,90 +150,12 @@ class CryptoListRepository {
         "history_price_last_30_days": 180.50,
         "history_price_last_50_days": 190.20,
       },
-      marketPriceUpnDown: {
-        "price_change_last_5_days": [
-          [0, 1],
-          [1.4, 2],
-          [2.4, 3.9],
-          [3, 6.5],
-          [4, 5],
-          [4.3, 8.1],
-          [5, 3.5],
-          [5.5, 4.5],
-          [6, 7.5],
-          [7.5, 10],
-          [8, 9.5],
-          [9, 12.2],
-          [9.3, 10.5],
-          [10, 12.4],
-        ],
-        "price_change_last_10_days": [
-          [0, 9],
-          [2.4, 3.9],
-          [4, 5],
-          [4.3, 5.1],
-          [5, 3.5],
-          [5.5, 5.5],
-          [6, 7.5],
-          [7.5, 6],
-          [8.5, 9.5],
-          [9, 10.2],
-          [9.3, 8.5],
-          [10, 9.4],
-        ],
-        "price_change_last_15_days": [
-          [0, 5],
-          [2.4, 3.9],
-          [4, 5],
-          [4.3, 5.1],
-          [5, 3.5],
-          [5.5, 5.5],
-          [6, 7.5],
-          [7.5, 6],
-          [8.5, 9.5],
-          [9, 10.2],
-          [9.3, 10.5],
-          [10, 9.4],
-        ],
-        "price_change_last_30_days": [
-          [0, 3.4],
-          [1, 4.9],
-          [2, 5.8],
-          [3, 7],
-          [3.5, 8],
-          [4, 9.5],
-          [5, 5],
-          [5.3, 7.5],
-          [5.7, 9.5],
-          [6.6, 7.5],
-          [7, 11],
-          [8, 8],
-          [8.5, 10],
-          [9, 12.2],
-          [10, 5],
-        ],
-        "price_change_last_50_days": [
-          [0, 5],
-          [2.4, 5.9],
-          [3, 4],
-          [4, 5],
-          [4.3, 5.1],
-          [5, 2],
-          [5.5, 5.5],
-          [6, 1.5],
-          [7, 10],
-          [8, 8],
-          [8.5, 10],
-          [8.8, 11],
-          [9, 10.5],
-          [10, 12.4],
-        ],
-      },
+      marketPriceUpnDown: marketPriceUpandDown,
     ),
   ];
 }
 
-const Map<String, List<List<double>>> MarketPriceUpandDown = {
+const Map<String, List<List<double>>> marketPriceUpandDown = {
   "price_change_last_5_days": [
     [0, 4],
     [1.4, 2],
