+++ lib/transactions/transactions_page.dart
@@ -1,4 +1,3 @@
-import 'package:crypto/shared/constants/app_colors.dart';
 import 'package:crypto/shared/constants/app_text_styles.dart';
 import 'package:flutter/material.dart';
